Revert "Merge master into branch to bring up to date."

+++ main/Camera.h
@@ -5,14 +5,14 @@
 
 struct cameraActiveDevice
 {
-	uint64_t ID;
-	uint64_t DevSceneRowID;
+	unsigned long long ID;
+	unsigned long long DevSceneRowID;
 	unsigned char DevSceneType;
 };
 
 struct cameraDevice
 {
-	uint64_t ID;
+	unsigned long long ID;
 	std::string Name;
     std::string Address;
 	std::string Username;
@@ -30,19 +30,19 @@ public:
 
     void ReloadCameras();
 
-	bool TakeSnapshot(const uint64_t CamID, std::vector<unsigned char> &camimage);
+	bool TakeSnapshot(const unsigned long long CamID, std::vector<unsigned char> &camimage);
 	bool TakeSnapshot(const std::string &CamID, std::vector<unsigned char> &camimage);
 	bool TakeRaspberrySnapshot(std::vector<unsigned char> &camimage);
 	bool TakeUVCSnapshot(const std::string &device, std::vector<unsigned char> &camimage);
-	cameraDevice* GetCamera(const uint64_t CamID);
+	cameraDevice* GetCamera(const unsigned long long CamID);
 	cameraDevice* GetCamera(const std::string &CamID);
-	uint64_t IsDevSceneInCamera(const unsigned char DevSceneType, const uint64_t DevSceneID);
-	uint64_t IsDevSceneInCamera(const unsigned char DevSceneType, const std::string &DevSceneID);
+	unsigned long long IsDevSceneInCamera(const unsigned char DevSceneType, const unsigned long long DevSceneID);
+	unsigned long long IsDevSceneInCamera(const unsigned char DevSceneType, const std::string &DevSceneID);
 
 	bool EmailCameraSnapshot(const std::string &CamIdx, const std::string &subject);
 	std::string GetCameraURL(cameraDevice *pCamera);
 	std::string GetCameraURL(const std::string &CamID);
-	std::string GetCameraURL(const uint64_t CamID);
+	std::string GetCameraURL(const unsigned long long CamID);
 private:
 	void ReloadCameraActiveDevices(const std::string &CamID);
